Cleaning the repository
Cleaning the repository
