+++ .gitignore
@@ -8,24 +8,10 @@ Environments/MultiTaxiEnv/__pycache__/
 Environments/SingleTaxiEnv/__pycache__/
 Environments/__pycache__/
 Transforms/__pycache__/
-Transforms/ckpt/
 Transforms/taxi_example_data/all_single_transformed_envs.pkl
-Agents/RL_agents/saved_models/
-Agents/TrainedAgents/trained_models/
-Transforms/taxi_example_data/taxi_transformed_env/
-Testing/ckpt/
-ckpt/
-Transforms/taxi_example_data/allowed_features.pkl
-Transforms/taxi_example_data/allowed_states.pkl
-Transforms/taxi_example_data/diff_dict.pkl
-Transforms/taxi_example_data/next_state_by_action.pkl
-Transforms/taxi_example_data/not_allowed_features.pkl
-Transforms/taxi_example_data/not_allowed_features_with_duplicates.pkl
-Transforms/taxi_example_data/preconditions.pkl
-Transforms/taxi_example_data/state_by_diff.pkl
-Transforms/taxi_example_data/not_allowed_states.pkl
-results/
-results/results_500_agents_on_transformed_envs.pkl
-Testing/
+Agents/TrainedAgents/trained_models_on_small_taxi_env/
+Agents/TrainedAgents/trained_models_on_big_taxi_env/
+Transforms/taxi_example_data/small_taxi_transformed_env/
 Transforms/taxi_example_data/big_taxi_transformed_env/
-Transforms/taxi_transform_search_env_different_reward.pkl
+results/
+Agents/TrainedAgents/results/
