--- Agents/TrainedAgents/trained_models/keras_dqn_0_(4,)_[0]_2_(4,)_[0]/checkpoint
@@ -1,2 +1,0 @@
-model_checkpoint_path: "keras_dqn_0_(4,)_[0]_2_(4,)_[0]"
-all_model_checkpoint_paths: "keras_dqn_0_(4,)_[0]_2_(4,)_[0]"
