--- Agents/TrainedAgents/trained_models/keras_dqn_original_env/checkpoint
@@ -1,2 +1,0 @@
-model_checkpoint_path: "dqn_model_on_original_env"
-all_model_checkpoint_paths: "dqn_model_on_original_env"
