+++ Transforms/env_precinditions.py
@@ -1,11 +1,10 @@
 import copy
 import pickle
-
 import pandas as pd
 import numpy as np
 from itertools import combinations
 
-from Environments.SingleTaxiEnv.single_taxi_env import SingleTaxiEnv
+# from Environments.SingleTaxiEnv.single_taxi_env import SingleTaxiEnv
 
 
 def change_representation(state_info_i):
