+++ create_single_taxi_transforms.py
@@ -1,21 +1,12 @@
 import multiprocessing
-import pickle
-
-from Agents.RL_agents import rl_agent
 from Agents.RL_agents.q_learning_agents import KERAS_DQN
 from Transforms.single_taxi_transforms import *
 from constants import *
-from utils import get_env
-
-AGENT_DATA_PATH = "Agents/TrainedAgents/"
-TRAINED_AGENTS_DIR_PATH = AGENT_DATA_PATH + "trained_models/"
-ENV_DIR_PATH = "Transforms/taxi_example_data/taxi_transformed_env/single_transform_envs/"
-RESULTS_DIR_PATH = AGENT_DATA_PATH + "results/"
 
 
 def check_success_rate():
     success_rates = {}
-    for file in os.listdir(ENV_DIR_PATH):
+    for file in os.listdir(TRAINED_AGENT_ON_SPECIFIC_TRANSFORM_NUM_PATH):
         success_rate = visualize_trained_agent(file)
         success_rates[file[:-4]] = success_rate
 
@@ -26,29 +17,28 @@ def check_success_rate():
 
 def visualize_trained_agent(file_name):
     transform_name = file_name[:-4] if file_name[-4:] == ".pkl" else file_name
-    file = open(ENV_DIR_PATH + file_name, "rb")
+    file = open(TRAINED_AGENT_ON_SPECIFIC_TRANSFORM_NUM_PATH + file_name, "rb")
     new_env = pickle.load(file)
 
     agent_name = KERAS_DQN
     model_name = agent_name + '_' + transform_name
     dir_path = TRAINED_AGENTS_DIR_PATH + model_name + '/'
-    new_agent = rl_agent.create_agent(new_env, agent_name)
-    new_agent = new_agent.load_existing_agent(dir_path + model_name)
+    new_agent = rl_agent.create_agent(new_env, agent_name)  # TODO - to fix this two lines
+    new_agent = load_existing_agent(new_env, agent_name, transform_name, dir_path + model_name)
     evaluation_result = rl_agent.run(new_agent, 5, method=EVALUATE)
 
     result_name = transform_name + "_" + agent_name
     result_dir = result_name + "_200000/"
-    file = open(RESULTS_DIR_PATH + result_dir + result_name + "_explanation.pkl", "rb")
+    file = open(TRAINED_AGENT_RESULTS_PATH + result_dir + result_name + "_explanation.pkl", "rb")
     explanation = pickle.load(file)
-    file = open(RESULTS_DIR_PATH + result_dir + result_name + "_result.pkl", "rb")
+    file = open(TRAINED_AGENT_RESULTS_PATH + result_dir + result_name + "_result.pkl", "rb")
     result = pickle.load(file)
 
     return result[transform_name][SUCCESS_RATE]
 
 
 def create_single_taxi_transforms():
-    a_file = open("Transforms/taxi_example_data/" + PRECONDITIONS_FILE_NAME, "rb")
-    cur_env_preconditions = pickle.load(a_file)
+    cur_env_preconditions = load_pkl_file(PRECONDITIONS_PATH)
     processes = []
     functions = [generate_single_transforms, generate_double_of_transforms, generate_triple_of_transforms]
     for func in functions:
@@ -59,11 +49,10 @@ def create_single_taxi_transforms():
     for p in processes:
         p.join()
 
-
-if __name__ == '__main__':
-    # check_success_rate()
-    # create_single_taxi_transforms()
-    a_file = open("Transforms/taxi_example_data/" + PRECONDITIONS_FILE_NAME, "rb")
-    cur_env_preconditions = pickle.load(a_file)
-    generate_triple_of_transforms(cur_env_preconditions)
-    print("DONE!")
+# if __name__ == '__main__':
+# check_success_rate()
+# create_single_taxi_transforms()
+# a_file = open("Transforms/taxi_example_data/" + PRECONDITIONS_FILE_NAME, "rb")
+# cur_env_preconditions = pickle.load(a_file)
+# generate_triple_of_transforms(cur_env_preconditions)
+# print("DONE!")
