+++ visualize.py
@@ -1,5 +1,5 @@
 import matplotlib.pyplot as plt
-# from qbstyles import mpl_style
+import numpy as np
 from constants import *
 
 # mpl_style(dark=False, minor_ticks=True)
